large changes to export all structs and functions to a shared library, installed into /usr/local/lib and /usr/local/include/concluster. Removed examples directory and tests related. Added seperate build system for the tests, which now rely on the library being installed.
