+++ include/cluster.h
@@ -1,6 +1,8 @@
 #ifndef CLUSTER_H
 #define CLUSTER_H
 
+#define EXPORT __attribute__((__visibility__("default")))
+
 #include "node.h"
 
 /**
@@ -9,7 +11,7 @@
  * Additional cluster-level logic should be implemented in a wrapper.
  * For creation/deletion information on for the node array, see node.h
  */
-typedef struct {
+EXPORT typedef struct {
     ccon_node** nodes;
     int node_count;
 } ccon_cluster;
@@ -23,7 +25,7 @@ typedef struct {
  * @param node_count 
  * @return cluster* struct or NULL.
  */
-ccon_cluster* ccon_create_cluster(ccon_node** nodes, int node_count);
+EXPORT ccon_cluster* ccon_create_cluster(ccon_node** nodes, int node_count);
 
 /**
  * @brief Create a cluster struct, cloning a node_struct struct node_count times.
@@ -34,7 +36,7 @@ ccon_cluster* ccon_create_cluster(ccon_node** nodes, int node_count);
  * @param node_count number of nodes to create.
  * @return cluster* cluster struct or NULL.
  */
-ccon_cluster* ccon_create_cluster_from_node(ccon_node* node_struct, int node_count);
+EXPORT ccon_cluster* ccon_create_cluster_from_node(ccon_node* node_struct, int node_count);
 
 /**
  * @brief Create a cluster of population node_count from default ccon_node struct.
@@ -44,7 +46,7 @@ ccon_cluster* ccon_create_cluster_from_node(ccon_node* node_struct, int node_cou
  * @param node_count number of nodes to have in struct.
  * @return ccon_cluster* ccon_cluster struct or NULL.
  */
-ccon_cluster* ccon_create_cluster_from_default_node(int node_count);
+EXPORT ccon_cluster* ccon_create_cluster_from_default_node(int node_count);
 
 /**
  * @brief deletes a ccon_cluster struct.
@@ -54,7 +56,7 @@ ccon_cluster* ccon_create_cluster_from_default_node(int node_count);
  * @param cluster struct to free.
  * @return int 0 on success, -1 on failure.
  */
-int ccon_delete_cluster(ccon_cluster** cluster);
+EXPORT int ccon_delete_cluster(ccon_cluster** cluster);
 
 /**
  * @brief delete ccon_node from ccon_cluster struct.
@@ -64,7 +66,7 @@ int ccon_delete_cluster(ccon_cluster** cluster);
  * @param node_index index of struct in ccon_cluster
  * @return int 0 on success, -1 on failure
  */
-int ccon_delete_cluster_node(ccon_cluster* cluster, int node_index);
+EXPORT int ccon_delete_cluster_node(ccon_cluster* cluster, int node_index);
 
 /**
  * @brief edit a ccon_node in-place in a cluster.
@@ -84,7 +86,7 @@ int ccon_delete_cluster_node(ccon_cluster* cluster, int node_index);
  * @param deep_copy 
  * @return int 0 on success, -1 on failure.
  */
-int ccon_edit_cluster_node(ccon_cluster* cluster,
+EXPORT int ccon_edit_cluster_node(ccon_cluster* cluster,
                            int node_index,
                            ccon_n_node_id* id,
                            ccon_n_node_role* role,
@@ -107,6 +109,6 @@ int ccon_edit_cluster_node(ccon_cluster* cluster,
  * @param node_index 
  * @return int 0 on success, -1 on fail.
  */
-int ccon_insert_cluster_node(ccon_cluster** cluster, ccon_node* node, int node_index);
+EXPORT int ccon_insert_cluster_node(ccon_cluster** cluster, ccon_node* node, int node_index);
 
 #endif
