+++ include/single_server.h
@@ -1,6 +1,8 @@
 #ifndef SINGLE_SERVERS_H
 #define SINGLE_SERVERS_H
 
+#define EXPORT __attribute__((__visibility__("default")))
+
 #include "address.h"
 
 /**
@@ -17,7 +19,7 @@
  *      ccon_n_node_address -> set of message bus channels
  *      serve -> subscribe function
  */
-typedef struct {
+EXPORT typedef struct {
     ccon_n_node_address* address;
     int invocation_status;
     void* (*serve) (void*);
@@ -32,7 +34,7 @@ typedef struct {
  * @param serve void* (void*) function pointer to represent a serve mechanism.
  * @return ccon_n_node_single_server* struct or NULL.
  */
-ccon_n_node_single_server* ccon_n_create_single_server(ccon_n_node_address* address, void* (*serve) (void*));
+EXPORT ccon_n_node_single_server* ccon_n_create_single_server(ccon_n_node_address* address, void* (*serve) (void*));
 
 /**
  * @brief performs deep copy on ccon_n_single_server struct to create new node_single_server.
@@ -41,7 +43,7 @@ ccon_n_node_single_server* ccon_n_create_single_server(ccon_n_node_address* addr
  * @param single_server struct to copy
  * @return ccon_n_node_single_server* struct, or NULL
  */
-ccon_n_node_single_server* ccon_n_copy_single_server(ccon_n_node_single_server* single_server);
+EXPORT ccon_n_node_single_server* ccon_n_copy_single_server(ccon_n_node_single_server* single_server);
 
 /**
  * @brief deletes a ccon_n_node_single_server struct.
@@ -51,6 +53,6 @@ ccon_n_node_single_server* ccon_n_copy_single_server(ccon_n_node_single_server* 
  * @param node_single_server struct to free.
  * @return int 0 on success, -1 on failure.
  */
-int ccon_n_delete_single_server(ccon_n_node_single_server** single_server);
+EXPORT int ccon_n_delete_single_server(ccon_n_node_single_server** single_server);
 
 #endif
