+++ src/address.c
@@ -1,4 +1,4 @@
-#include "../../include/node/address.h"
+#include "include/address.h"
 
 #include <stdlib.h>
 #include <string.h>
