+++ src/servers.c
@@ -1,4 +1,4 @@
-#include "../../include/node/servers.h"
+#include "include/servers.h"
 
 #include <stdlib.h>
 #include <memory.h>
