+++ src/single_server.c
@@ -1,4 +1,4 @@
-#include "../../include/node/single_server.h"
+#include "include/single_server.h"
 
 #include <stdlib.h>
 #include <memory.h>
